Separate the ThinEdgeJson data-model and the builder (visitor) from the JSON parser
* In-memory representation of ThinEdge JSON is in `data.rs`.

* ThinEdge JSON parser is in `json.rs` (and `stream.rs`).

* A `MeasurementVisitor` that builds up in-memory `ThinEdgeJson` is in
  `builder.rs` (I don't like the name as it can be confused with the
  builder pattern!).

* Do validation of the f64 measurement values at parse-time and not in
  the builder. The builder will expect valid measurement data being
  passed. Iideally we'd wrap that in a newtype `struct
  MeasurementValue(f64)` to ensure that contract at the type-level.
