+++ mapper/thin_edge_json/src/lib.rs
@@ -1,6 +1,8 @@
 //! A library to create [ThinEdgeJson][1] from bytes of json data by validating it.
 //! [1]: https://github.com/thin-edge/thin-edge.io/blob/main/docs/src/architecture/thin-edge-json.md
 
+pub mod builder;
+pub mod data;
 pub mod group;
 pub mod json;
 pub mod measurement;
